requirements.txt
asgiref==3.5.0
autocorrect==2.6.1
blis==0.7.6
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cycler==0.11.0
cymem==2.0.6
Django==4.0.3
djangorestframework==3.13.1
fonttools==4.30.0
idna==3.3
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
langcodes==3.3.0
MarkupSafe==2.1.0
matplotlib==3.5.1
murmurhash==1.0.6
nltk==3.7
numpy==1.22.3
packaging==21.3
pandas==1.4.1
pathy==0.6.1
Pillow==9.0.1
preshed==3.0.6
pydantic==1.8.2
pyenchant==3.2.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
regex==2022.3.2
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
smart-open==5.2.1
spacy==3.2.3
spacy-legacy==3.0.9
spacy-loggers==1.0.1
sqlparse==0.4.2
srsly==2.4.2
thinc==8.0.13
threadpoolctl==3.1.0
tqdm==4.63.0
typer==0.4.0
typing_extensions==4.1.1
tzdata==2021.5
urllib3==1.26.8
wasabi==0.9.0
